Literally just reuse ofstream object. Also format all the code.

+++ WorldChunkToolCPP/ChunkDecrypter.h
@@ -9,8 +9,8 @@
 class ChunkDecrypter
 {
 public:
-	ChunkDecrypter(const std::string& chunkKeyPatternFilePath=CHUNK_KEY_SEQUENCE_FILE_NAME);
-	static void DecryptChunk(uint8_t* data, const int length, const int dictionaryCount);
+	ChunkDecrypter(const std::string& chunkKeyPatternFilePath = CHUNK_KEY_SEQUENCE_FILE_NAME);
+	static void DecryptChunk(uint8_t* data, const int length, const size_t dictionaryCount);
 private:
 	static std::vector<uint8_t> chunkKeyPattern;
 	static std::vector<std::vector<uint8_t>> chunkKeys;
